avoid RESTEasy server + fix Retrofit usage
The Apicurio Registry 1.x client libraries use JAX-RS Client and
Retrofit as HTTP clients. For JAX-RS Client, we need RESTEasy,
but not the full RESTEasy server; the client is enough. Unfortunately
Quarkus doesn't have an extension for pure RESTEasy Client, so here
we use `quarkus-rest-client`, which brings MicroProfile RestClient.
Compared to bringing full RESTEasy server, this is neglibile.

Retrofit works standalone and without an issue on JVM, but requires
some substitutions for native compilation. Previously, we used to
depend on a Camel Quarkus extension for this, but that depends on
an older Retrofit version (2.5.0). Apicurio Registry 1.x libraries
depend on a newer Retrofit version (2.9.0), so here we remove
the dependency on Camel Quarkus and implement the substitutions
for Retrofit ourselves. Compared to the Camel Quarkus extension,
the substitutions here are much smaller.

All in all, with these changes, using the Apicurio Registry 1.x
client libraries in Quarkus native projects become much easier.